% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{align_platform_columns}
\alias{align_platform_columns}
\title{Align platform columns}
\usage{
align_platform_columns(cor_dt, eset1, eset2)
}
\arguments{
\item{cor_dt}{correlation data.table}

\item{eset1}{eset1}

\item{eset2}{eset2}
}
\description{
Ensure: col1 = eset1 features, col2 = eset2 features
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   cor_dt <- subramanian.2016::top.cor.exiqon.metabolon
   exiqon    <- subramanian.2016::exiqon    \%>\%
                autonomics.import::filter_samples(condition==unique(.$condition)[2])
   metabolon <- subramanian.2016::metabolon \%>\%
                autonomics.import::filter_samples(condition==unique(.$condition)[2])
   cor_dt
   cor_dt \%>\% autonomics.integrate::align_platform_columns(exiqon, metabolon)
   cor_dt \%>\% autonomics.integrate::align_platform_columns(metabolon, exiqon)
}
align_platform_columns
}
