% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{annotate_correlations}
\alias{annotate_correlations}
\title{Annotate correlations}
\usage{
annotate_correlations(
  cor_dt,
  eset1,
  eset2,
  fvars1 = character(0),
  fvars2 = character(0)
)
}
\arguments{
\item{cor_dt}{correlation datatable returned by correlate_features}

\item{eset1}{eset1}

\item{eset2}{eset2}

\item{fvars1}{eset1 fvars}

\item{fvars2}{eset2 fvars}
}
\value{
correlations datatable with annotations
}
\description{
Annotate correlations
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   eset1 <- subramanian.2016::exiqon \%>\%
            autonomics.import::filter_samples(condition==unique(.$condition)[2])
   eset2 <- subramanian.2016::metabolon \%>\%
            autonomics.import::filter_samples(condition==unique(.$condition)[2]) \%>\%
            magrittr::extract(rowSums(!is.na(autonomics.import::exprs(.))) == ncol(.))
   cor_dt <- autonomics.integrate::correlate_features(eset1, eset2)
   cor_dt \%<>\% autonomics.integrate::annotate_correlations(
                  eset1, eset2, character(0), c('BIOCHEMICAL', 'SUB_PATHWAY'))
}
}
