% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate_features}
\alias{correlate_features}
\title{Correlate eset features}
\usage{
correlate_features(eset1, eset2)
}
\arguments{
\item{eset1}{eset}

\item{eset2}{eset}
}
\description{
Compute pairwise correlations between each featrue of eset1 and eset2.
}
\details{
Both esets must contain an fvar 'feature_id' and an svar 'sample_id'.
Then, correlations will be computed for each pair of feature_id from eset1 and eset2.
Samples between the two esets will be matched on sample_id.
}
\note{
Avoid the use of devtools::load_all in combination with this function,
      as a devtools bug prevents data.table (which is used in this function) from being
      loaded properly:
      https://stackoverflow.com/questions/24501245/data-table-throws-object-not-found-error
      Upon regular installation with devtools::install or R CMD INSTALL, this function works fine.
}
\examples{
if (require(subramanian.2016)){
   require(magrittr)
   exiqon <- subramanian.2016::exiqon \%>\%
             autonomics.import::filter_samples(condition==unique(.$condition)[2]) \%>\%
             magrittr::extract(1:10, )
   metabolon <- subramanian.2016::metabolon \%>\%
                autonomics.import::filter_samples(condition==unique(.$condition)[2]) \%>\%
                magrittr::extract(rowSums(!is.na(autonomics.import::exprs(.))) == ncol(.), ) \%>\%
                magrittr::extract(1:10, )
   rnaseq <- subramanian.2016::rnaseq \%>\%
             autonomics.import::filter_samples(condition==unique(.$condition)[2]) \%>\%
             magrittr::extract(1:10, )
   autonomics.integrate::correlate_features(exiqon, metabolon)
   autonomics.integrate::correlate_features(exiqon, rnaseq)
   autonomics.integrate::correlate_features(rnaseq, metabolon)
}
}
