% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_corplot_fvars.R
\name{default_corplot_fvars}
\alias{default_corplot_fvars}
\title{Default correlation plot fvars}
\usage{
default_corplot_fvars(object)
}
\arguments{
\item{object}{exprs object}
}
\description{
Default correlation plot fvars
}
