% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{order_correlations}
\alias{order_correlations}
\title{Re-order datatable on key vars}
\usage{
order_correlations(cor_dt, key_vars)
}
\arguments{
\item{cor_dt}{correlation datatable}

\item{key_vars}{key variables}
}
\value{
correlation datable, ordered on variable of interest
}
\description{
The columns are reshuffled such that that the key variables come first.
The rows are ordered on the first key variable
}
\examples{
if (require(subramanian.2016)){
   require(magrittr)
   eset1  <- subramanian.2016::exiqon
   eset2  <- subramanian.2016::metabolon
   cor_dt <- subramanian.2016::top.cor.exiqon.metabolon
   cor_dt \%<>\% autonomics.integrate::annotate_correlations(
                  eset1, eset2, fvars2 = c('BIOCHEMICAL', 'SUB_PATHWAY'))
   cor_dt \%<>\% autonomics.integrate::order_correlations('BIOCHEMICAL')
}
}
