% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Plot single page with correlation plots}
\usage{
plot_correlations(cor_dt, eset1, eset2,
  fvars1 = autonomics.integrate::default_corplot_fvars(eset1),
  fvars2 = autonomics.integrate::default_corplot_fvars(eset2),
  color_var = "subgroup", shape_var = NULL, xlab = NULL, ylab = NULL,
  ncol = 3)
}
\arguments{
\item{cor_dt}{correlation datatable}

\item{eset1}{first eset}

\item{eset2}{second eset}

\item{fvars1}{eset1 fvars mapped to panel annotation}

\item{fvars2}{eset2 fvars mapped to panel annotation}

\item{color_var}{eset1 svar mapped to color}

\item{shape_var}{eset1 svar mapped to shape}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{ncol}{number of columns}
}
\description{
Plot single page with correlation plots
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   cor_dt    <- subramanian.2016::top.cor.exiqon.metabolon
   exiqon    <- subramanian.2016::exiqon    \%>\%
                autonomics.import::filter_samples(condition==unique(.$condition)[2])
   metabolon <- subramanian.2016::metabolon \%>\%
                autonomics.import::filter_samples(condition==unique(.$condition)[2])
   autonomics.integrate::plot_correlations(cor_dt[1:10, ], exiqon, metabolon)
}
}
