% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations_to_file}
\alias{plot_correlations_to_file}
\title{Plot correlations}
\usage{
plot_correlations_to_file(
  cor_dt,
  eset1,
  eset2,
  file,
  fvars1 = autonomics.integrate::default_corplot_fvars(eset1),
  fvars2 = autonomics.integrate::default_corplot_fvars(eset2),
  color_var = "subgroup",
  shape_var = NULL,
  xlab = NULL,
  ylab = NULL,
  nrow = 3,
  ncol = 6,
  max_pages = Inf
)
}
\arguments{
\item{cor_dt}{correlation datable, as returned by autonomics.integrate::correlate_features}

\item{eset1}{eset1 first eset}

\item{eset2}{eset2 second eset}

\item{file}{file to which to print plots}

\item{fvars1}{eset1 fvars used for faceting}

\item{fvars2}{eset2 fvars used for faceting}

\item{color_var}{eset1 svar mapped to color}

\item{shape_var}{eset1 svar mapped to shape}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{nrow}{no of rows on one page}

\item{ncol}{no of columns on one page}

\item{max_pages}{max no of pages}
}
\description{
Plot correlations
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   cor_dt    <- subramanian.2016::top.cor.exiqon.metabolon
   exiqon    <- subramanian.2016::exiqon    \%>\%
                autonomics.import::filter_samples(condition==unique(.$condition)[2])
   metabolon <- subramanian.2016::metabolon \%>\%
                autonomics.import::filter_samples(condition==unique(.$condition)[2])
   file <- tempfile() \%>\% paste0('.pdf') \%T>\% message()
   autonomics.integrate::plot_correlations_to_file(cor_dt, exiqon, metabolon, file = file)
   autonomics.integrate::plot_correlations_to_file(cor_dt, exiqon, metabolon, file = file,
                                           fvars2 = c('BIOCHEMICAL', 'SUB_PATHWAY'))
}
}
