% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_correlations.R
\name{write_correlations}
\alias{write_correlations}
\title{Write correlations to file}
\usage{
write_correlations(
  cor.dt,
  eset1,
  eset2,
  file,
  fvars1 = character(0),
  fvars2 = character(0)
)
}
\arguments{
\item{cor.dt}{correlation datable}

\item{eset1}{eset}

\item{eset2}{eset}

\item{file}{file}

\item{fvars1}{eset1 fvars}

\item{fvars2}{eset2 fvars}
}
\description{
Write correlations to file
}
\details{
Note that fvars are included in the file, but feature ids not
(except if fvars are empty).
}
\examples{
if (require(subramanian.2016)){
   cor.dt <- subramanian.2016::top.cor.exiqon.metabolon
   eset1  <- subramanian.2016::exiqon
   eset2  <-  subramanian.2016::metabolon
   fvars1 <- character(0)
   fvars2 <- c('BIOCHEMICAL', 'SUB_PATHWAY')
}
}
